+++ src/components/Main/Metrics/MetricsContent/MetricsContent.jsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from "react";
+import { useCallback, useContext, useEffect, useState } from "react";
 import MetricsItem from "./MetricsItem/MetricsItem.jsx";
 import $MetricsContent from "./MetricsContent.styles.jsx";
 import { apiCallForMetrics, MetricsContext } from "../../../../store/metrics-context.jsx";
@@ -7,21 +7,30 @@ export default function MetricsContent() {
   const { currentStep } = useContext(MetricsContext);
   const [availableMetrics, setAvailableMetrics] = useState([]);
 
-  useEffect(() => {  
-    const metrics = apiCallForMetrics(currentStep, 0);
-    setAvailableMetrics((_prevMetrics) => metrics);
-  }, [currentStep])
+  useEffect(() => {
+    async function fetchData() {
+      try {
+        await apiCallForMetrics(currentStep).then(resData => {
+          setAvailableMetrics((_prevMetrics) => resData);
+        });        
+      } catch (error) {
+        //TODO
+      }
+    }
 
-  function sessionStorageStepHandler(itemId, itemName) {
+    fetchData()
+  }, [currentStep]);
+
+  const sessionStorageStepHandler = useCallback((itemId, itemName) => {
     let stepItemInfo = JSON.parse(sessionStorage.getItem("stepItemInfo"));
-    if (stepItemInfo === undefined || stepItemInfo === null) {      
+    if (stepItemInfo === undefined || stepItemInfo === null) {     
       stepItemInfo = [{itemId: itemId, itemName: itemName}];
       sessionStorage.setItem("stepItemInfo", JSON.stringify(stepItemInfo))
     } else {
       stepItemInfo = [ ...stepItemInfo, {itemId: itemId, itemName: itemName}];
       sessionStorage.setItem("stepItemInfo", JSON.stringify(stepItemInfo));
     }
-  }
+  }, []);
 
   return (
     <$MetricsContent>