+++ src/components/Main/Metrics/MetricsContent/MetricsItem/MetricsItem.jsx
@@ -28,17 +28,17 @@ function descendantTooltipHandler(currentStep, role) {
     case 3:
       return STEPS[currentStep + 1].stepName + "s";
     case 4:
-      return role;
+      return role ? role.name : "";
   }
 }
 
 export default function MetricsItem({ metric, index, stepInfoHandler }) {
-  const { currentStep, stepHandler } = useContext(MetricsContext);
+  const { currentStep, stepHandler, thresholds } = useContext(MetricsContext);
   const [currentUsage, setCurrentUsage] = useState(0);
 
-  const { id, name, carbon_limit } = metric;
-  const usage = ((currentUsage / carbon_limit) * 100).toFixed(2) + " %";
-  const threshold = carbonBalance(id);
+  const { id, name, carbonLimit } = metric;
+  const usage = ((currentUsage / carbonLimit) * 100).toFixed(2) + " %";
+  const threshold = carbonBalance();
 
   useEffect(() => {
     // To DO strzal do bazy po current usage dla company/tribe etc
@@ -53,14 +53,13 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
     return 12;
   }
 
-  function carbonBalance(id) {
-    // TODO api call to get thresholds by id
-    // meanwhile mock
-    // threshold powinienb byc gdzies wyzej i moze w kontekscie albo storage
-    const returnedThresholds = [200, 100, 90, 0];
-    
+  function roleAsDescendant() {
+    return metric.role && metric.role.name.charAt(0);
+  }
+
+  function carbonBalance() {
     const balance = parseFloat(usage);
-    const threshold = returnedThresholds.findIndex(
+    const threshold = thresholds.findIndex(
       (element) => balance > element
     );
     return threshold;
@@ -97,7 +96,7 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
           <p>
             {currentStep < 4
               ? descendantsCount(id)
-              : metric.role && metric.role.charAt(0)}
+              : roleAsDescendant()}
           </p>
           <Tooltip id={"descendant_tooltip_" + index} />
         </$Descendants>
@@ -120,9 +119,9 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
         data-tooltip-delay-show={1000}
         data-tooltip-place={"bottom"}
       >
-        <div>{carbon_limit}</div>
+        <div>{carbonLimit}</div>
         <div>{currentUsage}</div>
-        <div>{carbon_limit - currentUsage}</div>
+        <div>{carbonLimit - currentUsage}</div>
         <Tooltip id={"carbon_tooltip_" + index} />
       </$Carbon>
     </$MetricsItem>