+++ src/components/Main/Metrics/MetricsEmployee/Dropdown/Dropdown.jsx
@@ -14,7 +14,7 @@ export default function Dropdown({ title, details, children }) {
   const [isOpen, setIsOpen] = useState(false);
 
   function toggleDropdown() {
-    setIsOpen((_prevIsOpen) => !isOpen);
+    setIsOpen((_prevIsOpen) => !_prevIsOpen);
   }
 
   return (
@@ -29,7 +29,6 @@ export default function Dropdown({ title, details, children }) {
             {isOpen ? <FiArrowDownCircle /> : <FiArrowRightCircle />}
           </$Icon>
         )}
-        {/* <$Icon>{isOpen ? <FiArrowDownCircle /> : <FiArrowRightCircle />}</$Icon> */}
         <$Title>{title}</$Title>
       </$Head>
       <CSSTransition