+++ src/components/Main/Metrics/MetricsStepper/MetricsStepper.styles.jsx
@@ -1,7 +1,7 @@
 import { styled } from "styled-components";
 
 export const $MetricsStepper = styled.div`
-  margin: 20px 100px 20px 0;
+  margin: 25px 100px 25px 0;
   display: flex;
   justify-content: space-evenly;
 `;